Pull saveContract and getModelEntities to util.
